refactor(tests): reorganize test helpers and mocks into config/tests directory
