+++ src/domain/auth/use-cases/__test__/register-user.test.ts
@@ -2,7 +2,7 @@ import { RegisterUser } from "../register-user";
 import { RegisterDto } from "@/domain/auth/dtos/register.dto";
 import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
 import { UserEntity } from "@/domain/user/entities/user.entity";
-import { MockAuthRepository } from "@/config/__tests__/mocks/auth-repository.mock";
+import { MockAuthRepository } from "@/config/tests/__mocks__/auth-repository.mock";
 
 describe("RegisterUser", () => {
   let registerUser: RegisterUser;
